+++ game/static/game/js/event.js
@@ -10,3 +10,9 @@ ocargo.Event.prototype.sendEvent = function(eventType, details) {
 $(document).ready(function() {
   ocargo.event = new ocargo.Event();
 });
+
+// The play button seems to have very complicated implementation
+// therefore, retriggerring the current event when click run code
+const runCode = () => {
+  $("#play_radio").click()
+}