+++ game/tests/test_level_editor.py
@@ -5,7 +5,6 @@ from common.tests.utils.classes import create_class_directly
 from common.tests.utils.organisation import create_organisation_directly
 from common.tests.utils.student import create_school_student_directly
 from common.tests.utils.teacher import signup_teacher_directly
-from game.models import Level
 from deploy import captcha
 from django.core import mail
 from django.test.client import Client
@@ -13,6 +12,7 @@ from django.test.testcases import TestCase
 from django.urls import reverse
 from hamcrest import assert_that, equal_to
 
+from game.models import Level
 from game.tests.utils.level import create_save_level
 from game.tests.utils.teacher import add_teacher_to_school, create_school
 
@@ -23,11 +23,7 @@ class LevelEditorTestCase(TestCase):
         "pythonEnabled": False,
         "decor": [],
         "blocklyEnabled": True,
-        "blocks": [
-            {"type": "move_forwards"},
-            {"type": "turn_left"},
-            {"type": "turn_right"},
-        ],
+        "blocks": [{"type": "move_forwards"}, {"type": "turn_left"}, {"type": "turn_right"}],
         "max_fuel": "50",
         "pythonViewEnabled": False,
         "character": "3",
@@ -35,8 +31,7 @@ class LevelEditorTestCase(TestCase):
         "theme": 1,
         "anonymous": False,
         "cows": "[]",
-        "path": '[{"coordinate":[3,5],"connectedNodes":[1]},{"coordinate":[3,4],'
-        '"connectedNodes":[0]}]',
+        "path": '[{"coordinate":[3,5],"connectedNodes":[1]},{"coordinate":[3,4],' '"connectedNodes":[0]}]',
         "traffic_lights": "[]",
         "destinations": "[[3,4]]",
     }
@@ -58,11 +53,7 @@ class LevelEditorTestCase(TestCase):
     def login(self, email, password):
         self.client.post(
             reverse("teacher_login"),
-            {
-                "auth-username": email,
-                "auth-password": password,
-                "teacher_login_view-current_step": "auth",
-            },
+            {"auth-username": email, "auth-password": password, "teacher_login_view-current_step": "auth"},
             follow=True,
         )
 
@@ -72,10 +63,7 @@ class LevelEditorTestCase(TestCase):
     def student_login(self, name, access_code, password):
         self.client.post(
             reverse("student_login", kwargs={"access_code": access_code}),
-            {
-                "username": name,
-                "password": password,
-            },
+            {"username": name, "password": password},
             follow=True,
         )
 
@@ -100,9 +88,7 @@ class LevelEditorTestCase(TestCase):
         response = self.client.post(url, {"data": json.dumps(self.LEVEL_DATA1)})
 
         assert_that(response.status_code, equal_to(200))
-        sharing_info1 = json.loads(
-            self.get_sharing_information(json.loads(response.content)["id"]).getvalue()
-        )
+        sharing_info1 = json.loads(self.get_sharing_information(json.loads(response.content)["id"]).getvalue())
         assert_that(sharing_info1["teacher"]["shared"], equal_to(True))
         assert_that(len(mail.outbox), equal_to(1))
 
@@ -119,11 +105,7 @@ class LevelEditorTestCase(TestCase):
             "pythonEnabled": False,
             "decor": [],
             "blocklyEnabled": True,
-            "blocks": [
-                {"type": "move_forwards"},
-                {"type": "turn_left"},
-                {"type": "turn_right"},
-            ],
+            "blocks": [{"type": "move_forwards"}, {"type": "turn_left"}, {"type": "turn_right"}],
             "max_fuel": "50",
             "pythonViewEnabled": False,
             "character": "3",
@@ -131,17 +113,14 @@ class LevelEditorTestCase(TestCase):
             "theme": 1,
             "anonymous": True,
             "cows": "[]",
-            "path": '[{"coordinate":[3,5],"connectedNodes":[1]},{"coordinate":[3,4],'
-            '"connectedNodes":[0]}]',
+            "path": '[{"coordinate":[3,5],"connectedNodes":[1]},{"coordinate":[3,4],' '"connectedNodes":[0]}]',
             "traffic_lights": "[]",
             "destinations": "[[3,4]]",
         }
         response = self.client.post(url, {"data": json.dumps(data1)})
 
         assert_that(response.status_code, equal_to(200))
-        sharing_info1 = json.loads(
-            self.get_sharing_information(json.loads(response.content)["id"]).getvalue()
-        )
+        sharing_info1 = json.loads(self.get_sharing_information(json.loads(response.content)["id"]).getvalue())
         assert_that(sharing_info1["teacher"]["shared"], equal_to(True))
         assert_that(len(mail.outbox), equal_to(0))
 
@@ -149,9 +128,9 @@ class LevelEditorTestCase(TestCase):
         email1, password1 = signup_teacher_directly()
         teacher1 = Teacher.objects.get(new_user__email=email1)
         self.login(email1, password1)
-        level_id = create_save_level(teacher1)
+        level = create_save_level(teacher1)
 
-        sharing_info1 = json.loads(self.get_sharing_information(level_id).getvalue())
+        sharing_info1 = json.loads(self.get_sharing_information(level.id).getvalue())
         assert_that(len(sharing_info1["teachers"]), equal_to(0))
 
     def test_level_sharing_with_school(self):
@@ -162,13 +141,13 @@ class LevelEditorTestCase(TestCase):
         teacher2 = Teacher.objects.get(new_user__email=email2)
 
         self.login(email1, password1)
-        level_id = create_save_level(teacher1)
+        level = create_save_level(teacher1)
 
         school1 = create_school()
         add_teacher_to_school(teacher1, school1)
         add_teacher_to_school(teacher2, school1)
 
-        sharing_info1 = json.loads(self.get_sharing_information(level_id).getvalue())
+        sharing_info1 = json.loads(self.get_sharing_information(level.id).getvalue())
         assert_that(len(sharing_info1["teachers"]), equal_to(1))
 
     def test_level_sharing_with_empty_school(self):
@@ -176,12 +155,12 @@ class LevelEditorTestCase(TestCase):
         teacher1 = Teacher.objects.get(new_user__email=email1)
 
         self.login(email1, password1)
-        level_id = create_save_level(teacher1)
+        level = create_save_level(teacher1)
 
         school1 = create_school()
         add_teacher_to_school(teacher1, school1)
 
-        sharing_info1 = json.loads(self.get_sharing_information(level_id).getvalue())
+        sharing_info1 = json.loads(self.get_sharing_information(level.id).getvalue())
         assert_that(len(sharing_info1["teachers"]), equal_to(0))
 
     def test_level_sharing_permissions(self):
@@ -192,8 +171,8 @@ class LevelEditorTestCase(TestCase):
         teacher2 = Teacher.objects.get(new_user__email=email2)
 
         self.login(email1, password1)
-        level_id = create_save_level(teacher1)
-        share_url = reverse("share_level_for_editor", args=[level_id])
+        level = create_save_level(teacher1)
+        share_url = reverse("share_level_for_editor", args=[level.id])
 
         school1 = create_school()
         add_teacher_to_school(teacher1, school1)
@@ -201,18 +180,13 @@ class LevelEditorTestCase(TestCase):
 
         # Create a class and student for the second teacher
         _, class_name2, access_code2 = create_class_directly(email2)
-        student_name2, student_password2, student2 = create_school_student_directly(
-            access_code2
-        )
+        student_name2, student_password2, student2 = create_school_student_directly(access_code2)
 
         self.logout()
         self.login(email2, password2)
 
         # Second teacher can't share the level as it's not been shared with them yet
-        response = self.client.post(
-            share_url,
-            {"recipientIDs[]": [student2.new_user.id], "action": ["share"]},
-        )
+        response = self.client.post(share_url, {"recipientIDs[]": [student2.new_user.id], "action": ["share"]})
         assert response.status_code == 403
 
         # Log in as the first teacher again
@@ -220,10 +194,7 @@ class LevelEditorTestCase(TestCase):
         self.login(email1, password1)
 
         # Share the level with the second teacher
-        response = self.client.post(
-            share_url,
-            {"recipientIDs[]": [teacher2.new_user.id], "action": ["share"]},
-        )
+        response = self.client.post(share_url, {"recipientIDs[]": [teacher2.new_user.id], "action": ["share"]})
         assert response.status_code == 200
 
         # Log in as the second teacher
@@ -231,13 +202,10 @@ class LevelEditorTestCase(TestCase):
         self.login(email2, password2)
 
         # Now the second teacher should be able to share the level
-        response = self.client.post(
-            share_url,
-            {"recipientIDs[]": [student2.new_user.id], "action": ["share"]},
-        )
+        response = self.client.post(share_url, {"recipientIDs[]": [student2.new_user.id], "action": ["share"]})
         assert response.status_code == 200
         # and load it
-        load_level_url = reverse("load_level_for_editor", kwargs={"levelID": level_id})
+        load_level_url = reverse("load_level_for_editor", kwargs={"levelID": level.id})
         response = self.client.get(load_level_url)
         assert response.status_code == 200
 
@@ -246,44 +214,31 @@ class LevelEditorTestCase(TestCase):
         self.student_login(student_name2, access_code2, student_password2)
 
         # Check that the student cannot share the level
-        sharing_info = json.loads(self.get_sharing_information(level_id).getvalue())
-        assert (
-            sharing_info["detail"]
-            == "You do not have permission to perform this action."
-        )
+        sharing_info = json.loads(self.get_sharing_information(level.id).getvalue())
+        assert sharing_info["detail"] == "You do not have permission to perform this action."
 
         # Check the student can view the level
-        response = self.client.get(reverse("play_custom_level", args=[level_id]))
+        response = self.client.get(reverse("play_custom_level", args=[level.id]))
         assert response.status_code == 200
 
     def test_level_can_only_be_edited_by_owner(self):
         email, password = signup_teacher_directly()
         create_organisation_directly(email)
         _, _, access_code = create_class_directly(email)
-        student_name, student_password, student = create_school_student_directly(
-            access_code
-        )
-        hacker_name, hacker_password, hacker = create_school_student_directly(
-            access_code
-        )
+        student_name, student_password, student = create_school_student_directly(access_code)
+        hacker_name, hacker_password, hacker = create_school_student_directly(access_code)
 
-        new_level = Level.objects.create(
-            owner_id=student.user_id, name="my_custom_level"
-        )
+        new_level = Level.objects.create(owner_id=student.user_id, name="my_custom_level")
 
         self.student_login(hacker_name, access_code, hacker_password)
-        resp = self.client.get(
-            reverse("level_editor_chosen_level", kwargs={"levelId": new_level.id})
-        )
+        resp = self.client.get(reverse("level_editor_chosen_level", kwargs={"levelId": new_level.id}))
 
         self.assertNotIn("level", resp.context)
 
         self.logout()
 
         self.student_login(student_name, access_code, student_password)
-        resp = self.client.get(
-            reverse("level_editor_chosen_level", kwargs={"levelId": new_level.id})
-        )
+        resp = self.client.get(reverse("level_editor_chosen_level", kwargs={"levelId": new_level.id}))
 
         self.assertIn("level", resp.context)
 
@@ -301,19 +256,14 @@ class LevelEditorTestCase(TestCase):
             "pythonEnabled": False,
             "decor": [],
             "blocklyEnabled": True,
-            "blocks": [
-                {"type": "move_forwards"},
-                {"type": "turn_left"},
-                {"type": "turn_right"},
-            ],
+            "blocks": [{"type": "move_forwards"}, {"type": "turn_left"}, {"type": "turn_right"}],
             "max_fuel": "50",
             "pythonViewEnabled": False,
             "name": "abc",
             "theme": 1,
             "anonymous": True,
             "cows": "[]",
-            "path": '[{"coordinate":[3,5],"connectedNodes":[1]},{"coordinate":[3,4],'
-            '"connectedNodes":[0]}]',
+            "path": '[{"coordinate":[3,5],"connectedNodes":[1]},{"coordinate":[3,4],' '"connectedNodes":[0]}]',
             "traffic_lights": "[]",
             "destinations": "[[3,4]]",
         }
@@ -329,8 +279,8 @@ class LevelEditorTestCase(TestCase):
         teacher1 = Teacher.objects.get(new_user__email=email1)
 
         self.login(email1, password1)
-        level_id = create_save_level(teacher1)
-        url = reverse("load_level_for_editor", kwargs={"levelID": level_id})
+        level = create_save_level(teacher1)
+        url = reverse("load_level_for_editor", kwargs={"levelID": level.id})
         response = self.client.get(url)
 
         assert_that(response.status_code, equal_to(200))
@@ -374,9 +324,7 @@ class LevelEditorTestCase(TestCase):
         email, password = signup_teacher_directly()
         create_organisation_directly(email)
         _, _, access_code = create_class_directly(email)
-        student_name, student_password, student = create_school_student_directly(
-            access_code
-        )
+        student_name, student_password, student = create_school_student_directly(access_code)
 
         teacher = Teacher.objects.get(new_user__email=email)